+++ sdk/rdbms/azure-mgmt-rdbms/azure/mgmt/rdbms/postgresql/operations/__init__.py
@@ -7,23 +7,41 @@
 # --------------------------------------------------------------------------
 
 from ._servers_operations import ServersOperations
+from ._replicas_operations import ReplicasOperations
 from ._firewall_rules_operations import FirewallRulesOperations
-from ._configurations_operations import ConfigurationsOperations
-from ._check_name_availability_operations import CheckNameAvailabilityOperations
-from ._location_based_capabilities_operations import LocationBasedCapabilitiesOperations
-from ._virtual_network_subnet_usage_operations import VirtualNetworkSubnetUsageOperations
-from ._operations import Operations
+from ._virtual_network_rules_operations import VirtualNetworkRulesOperations
 from ._databases_operations import DatabasesOperations
-from ._get_private_dns_zone_suffix_operations import GetPrivateDnsZoneSuffixOperations
+from ._configurations_operations import ConfigurationsOperations
+from ._server_parameters_operations import ServerParametersOperations
+from ._log_files_operations import LogFilesOperations
+from ._server_administrators_operations import ServerAdministratorsOperations
+from ._recoverable_servers_operations import RecoverableServersOperations
+from ._server_based_performance_tier_operations import ServerBasedPerformanceTierOperations
+from ._location_based_performance_tier_operations import LocationBasedPerformanceTierOperations
+from ._check_name_availability_operations import CheckNameAvailabilityOperations
+from ._operations import Operations
+from ._server_security_alert_policies_operations import ServerSecurityAlertPoliciesOperations
+from ._private_endpoint_connections_operations import PrivateEndpointConnectionsOperations
+from ._private_link_resources_operations import PrivateLinkResourcesOperations
+from ._server_keys_operations import ServerKeysOperations
 
 __all__ = [
     'ServersOperations',
+    'ReplicasOperations',
     'FirewallRulesOperations',
-    'ConfigurationsOperations',
-    'CheckNameAvailabilityOperations',
-    'LocationBasedCapabilitiesOperations',
-    'VirtualNetworkSubnetUsageOperations',
-    'Operations',
+    'VirtualNetworkRulesOperations',
     'DatabasesOperations',
-    'GetPrivateDnsZoneSuffixOperations',
+    'ConfigurationsOperations',
+    'ServerParametersOperations',
+    'LogFilesOperations',
+    'ServerAdministratorsOperations',
+    'RecoverableServersOperations',
+    'ServerBasedPerformanceTierOperations',
+    'LocationBasedPerformanceTierOperations',
+    'CheckNameAvailabilityOperations',
+    'Operations',
+    'ServerSecurityAlertPoliciesOperations',
+    'PrivateEndpointConnectionsOperations',
+    'PrivateLinkResourcesOperations',
+    'ServerKeysOperations',
 ]