+++ sdk/textanalytics/azure-ai-textanalytics/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.1.0b7 (Unreleased)
 
 **New Features**
+- Added enums `EntityConditionality`, `EntityCertainty`, and `EntityAssociation`.
 - Added `AnalyzeSentimentAction` as a supported action type for `begin_analyze_batch_actions`.
 
 ## 5.1.0b6 (2021-03-09)